requirements.txt
# Core ML libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==1.7.6

# MongoDB
pymongo==4.5.0

# Environment variables
python-dotenv==1.0.0

# Model serialization
joblib==1.3.2

# Visualization libraries
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1
kaleido==0.2.1

# Optional: For better performance
scipy==1.11.2
